+++ make_MagOS/files/patches/rootfs/MagOS-gdm.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 PFP=etc/X11/gdm/custom.conf
+[ -f $PFP ] || exit 0
 grep -q MagOS $PFP && exit 0
 cat > $PFP <<EOF
 #MagOS config for autologin
