+++ make_MagOS/files/patches/rootfs/MagOS-polipo.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
+#BUGFIX
+PFP=etc/rc.d/init.d/polipo
+[ -f $PFP ] || exit 0
+grep -q "\$remote_fs" $PFP && sed -i s/.\$remote_fs// $PFP
+grep -q "Default-Start:" $PFP || sed -i 's|^# Description:|# Default-Start: 2 3 4 5\n# Description:|' $PFP
+
 PFP=etc/polipo/config
+[ -f $PFP ] || exit 0
 grep -q MagOS $PFP && exit 0
 cp -p $PFP ${PFP}_default
 cat >$PFP <<EOF
