+++ make_MagOS/files/patches/rootfs/MagOS-tor.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 PFP=etc/tor/torrc
+[ -f $PFP ] || exit 0
 grep -q MagOS ${PFP} && exit 0
 cat >$PFP <<EOF
 #MagOS default config
